Attempt 2 on removing gunicorn from top-level CLI imports
